+++ src/parser/ELR/DFA/candidate.ts
@@ -111,6 +111,7 @@ export class Candidate<T> {
   tryReduce(
     buffer: readonly ASTNode<T>[],
     entryNTs: ReadonlySet<string>,
+    firstSets: ReadonlyMap<string, GrammarSet>,
     followSets: ReadonlyMap<string, GrammarSet>,
     lexer: ILexer,
     cascadeQueryPrefix: string | undefined,
@@ -122,6 +123,9 @@ export class Candidate<T> {
       matched: buffer.slice(-this.gr.rule.length),
       before: buffer.slice(0, -this.gr.rule.length),
       after: lexer.getRest(),
+      firstSets,
+      followSets,
+      gr: this.gr,
       lexer,
       $: (name) => {
         const result: ASTNode<T>[] = [];